GSSU/requirements.txt
pyautogui
pyperclip
pyyaml
google-generativeai
